Extract signing code into a separate module
